// apps/api/src/employee-forms/dto/create-employee-form-template.dto.d.ts
export declare class CreateEmployeeFormTemplateDto {
    businessId?: string;
    type?: string;
    title: string;
    description?: string;
    status?: string;
    version?: string;
    body?: string;
    fields?: string;
    fileUrl?: string;
    acknowledgementRequired?: boolean;
    requiresSignature?: boolean;
}